+++ frontend/soilhive-app/src/components/Map/SoilhiveMapSelectionToolbar.tsx
@@ -1,43 +1,46 @@
 /* eslint-disable */
+import SearchIcon from 'assets/icons/small-search-icon.svg?react';
 import ResetIcon from 'assets/icons/small-reset-icon.svg?react';
-import MarkerIcon from 'assets/icons/small-marker-icon.svg?react';
-import PlusIcon from 'assets/icons/small-plus-icon.svg?react';
+import CrossIcon from 'assets/icons/small-cross-icon.svg?react';
 import { Button } from '../UI';
 import { useTranslation } from 'react-i18next';
+import classNames from 'classnames';
+
+export type SoilhiveMapSelectionToolbarMode = 'drawing' | 'clear' | 'search';
 
 interface SoilhiveMapSelectionToolbarProps {
-  area?: number;
+  mode?: SoilhiveMapSelectionToolbarMode;
   onCancel: () => void;
   onReset: () => void;
-  onDrawAnother: () => void;
-  onShowResults: () => void;
 }
 
-export default function SoilhiveMapSelectionToolbar({ area = 0, onCancel, onReset, onDrawAnother, onShowResults }: SoilhiveMapSelectionToolbarProps) {
+export default function SoilhiveMapSelectionToolbar({ mode = 'drawing', onCancel, onReset }: SoilhiveMapSelectionToolbarProps) {
   const { t } = useTranslation('availability');
   return (
-    <div className="soilhive-map-selection-toolbar">
-      <Button size="tiny" type="tertiary" onClick={onReset} className="reset-button">
-        <ResetIcon />
-        {t('map_selection_toolbar.reset')}
-      </Button>
-      <Button size="tiny" type="tertiary" onClick={onCancel}>
-        {t('map_selection_toolbar.cancel')}
-      </Button>
-      <Button className="draw-another-button" size="tiny" type="secondary" onClick={onDrawAnother} isDisabled={true}>
-        <PlusIcon />
-        {t('map_selection_toolbar.draw_another_area')}
-      </Button>
-      <Button className="show-results-button" size="tiny" type="primary" onClick={onShowResults}>
-        <MarkerIcon />
-        {t('map_selection_toolbar.show_results')}
-      </Button>
-      <div className="selection-area">{area} km²</div>
-      <div className="break"></div>
-      <Button className="draw-another-button-mobile" size="tiny" type="secondary" onClick={onDrawAnother} isDisabled={true}>
-        <PlusIcon />
-        {t('map_selection_toolbar.draw_another_area')}
-      </Button>
+    <div className={classNames('soilhive-map-selection-toolbar', { drawing: mode === 'drawing' })}>
+      { mode === 'drawing' &&
+        <>
+          <Button size="tiny" type="tertiary" onClick={onReset} className="reset-button">
+            <ResetIcon />
+            {t('map_selection_toolbar.reset')}
+          </Button>
+          <Button size="tiny" type="secondary" onClick={onCancel}>
+            {t('map_selection_toolbar.cancel')}
+          </Button>
+        </>
+      }
+      { mode === 'clear' && 
+        <Button size="tiny" type="tertiary" onClick={onCancel} className="clear-button">
+          <CrossIcon />
+          {t('map_selection_toolbar.clear_selection')}
+        </Button>
+      }
+      { mode === 'search' && 
+        <Button size="tiny" type="tertiary" onClick={onCancel} className="search-area-button">
+            <SearchIcon />
+            {t('map_selection_toolbar.search_this_area')}
+          </Button>
+      }
     </div>
   );
 }